Auth_Service/auth_service/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
requests==2.31.0
django-extensions==3.2.3
django-redis

